Move test config to different packages.
