+++ hapi-fhir-storage-cr/src/test/java/ca/uhn/fhir/cr/dstu3/TestCrDstu3Config.java
@@ -1,9 +1,9 @@
 package ca.uhn.fhir.cr.dstu3;
 
 import ca.uhn.fhir.cr.TestHapiFhirCrPartitionConfig;
-import ca.uhn.fhir.cr.test.TestCqlProperties;
+import ca.uhn.fhir.cr.config.test.TestCqlProperties;
 import ca.uhn.fhir.cr.config.dstu3.CrDstu3Config;
-import ca.uhn.fhir.cr.test.TestCrConfig;
+import ca.uhn.fhir.cr.config.test.TestCrConfig;
 import org.cqframework.cql.cql2elm.CqlCompilerOptions;
 import org.cqframework.cql.cql2elm.model.CompiledLibrary;
 import org.cqframework.cql.cql2elm.model.Model;
